Reorganized components into separate modules.
