+++ hasher-matcher-actioner/webapp/src/Settings.js
@@ -1,0 +1,13 @@
+/**
+ * Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved
+ */
+
+import React from 'react';
+
+export default function Settings() {
+  return (
+    <>
+      <h1>Settings</h1>
+    </>
+  );
+}
